Surface: Start moving some code around, preparing to spilt into GL and Native implementations
