+++ Source/Fuse.Drawing.Surface/Android/AndroidGraphicsContext.java
@@ -14,12 +14,12 @@ public class AndroidGraphicsContext {
 
 	// dimensions of the framebuffer at any one point
 	public int width;
-	public int height; 
+	public int height;
 
 	// the texture ID given by a framebuffer
 	public int glTextureId;
 
-	public void saveCurrentMatrix() 
+	public void saveCurrentMatrix()
 	{
 		canvas.save();
 	}
@@ -27,23 +27,6 @@ public class AndroidGraphicsContext {
 	public void restoreCurrentMatrix()
 	{
 		canvas.restore();
-	}
-
-	// since we need to provide the linear gradient with
-	// different stops based on the rotation of the phone
-	// this class is used as a holder for all that until the 
-	// gradient is actually drawn
-	public static class LinearGradientStore
-	{
-		public int[] colors;
-		public float[] stops;
-
-
-		public String toString()
-		{
-			return (" " + colors
-				+ " " + stops);
-		}
 	}
 }
