+++ Source/Fuse.Drawing.Surface/Android/ISurfaceContext.java
@@ -1,0 +1,7 @@
+package com.fusetools.drawing.surface;
+
+import android.graphics.Canvas;
+
+public interface ISurfaceContext {
+	public Canvas getCanvas();
+}
